Add wildcard "*" support in permission setting.
Currently wildcard "*" is only allowed in permission checks but not in the permission settings. This allows permission setting to use wildcards also. If a permission setting is defined as `admin.posts.*` then the check `hasAccess('admin.posts.edit')` will return true.

This is especially helpful in L4 where it is possible to automatically do a check on every `controller.action` using a generic filter and no explicit checking anywhere else.